move bt_agent from app layer to hal layer, and run its GMainLoop in its own thread
