+++ bluetooth-speaker-app/app/CMakeLists.txt
@@ -18,7 +18,6 @@ target_link_libraries(433-bluetooth-speaker LINK_PRIVATE hal)
 target_link_libraries(433-bluetooth-speaker LINK_PRIVATE lcd)
 target_link_libraries(433-bluetooth-speaker LINK_PRIVATE ui)
 target_link_libraries(433-bluetooth-speaker LINK_PRIVATE lgpio)
-target_link_libraries(433-bluetooth-speaker LINK_PRIVATE Binc)
 
 # ALSA support
 find_package(ALSA REQUIRED)
