+++ bluetooth-speaker-app/hal/CMakeLists.txt
@@ -17,3 +17,5 @@ target_link_libraries(hal PUBLIC m)
 # Support GPIO
 find_library(GPIOD_LIBRARY gpiod)       # UNSURE IF NEEDED
 target_link_libraries(hal LINK_PRIVATE gpiod) # May need to change to HAL lib.
+
+target_link_libraries(hal LINK_PRIVATE Binc)
