+++ bluetooth-speaker-app/hal/include/hal/bt_agent.h
