+++ bluetooth-speaker-app/hal/src/bt_agent.c
@@ -1,4 +1,4 @@
-#include "app/bt_agent.h"
+#include "hal/bt_agent.h"
 
 #include <glib.h>
 #include <pthread.h>
@@ -17,13 +17,15 @@
 #define A2DP_PROF_UUID "0000110d-0000-1000-8000-00805f9b34fb"
 #define AVCRP_PROF_UUID "0000110e-0000-1000-8000-00805f9b34fb"
 
-GMainLoop *loop = NULL;
+GDBusConnection *dbusconn;
+GMainLoop *main_loop = NULL;
 Adapter *default_adapter = NULL;
-Advertisement *advertisement = NULL;
 Application *app = NULL;
 Agent *agent = NULL;
 
-void *agent_task(void *arg);
+pthread_t agent_thread_id;
+
+static void *agent_routine(void *arg);
 
 void on_central_state_changed(Adapter *adapter, Device *device)
 {
@@ -46,62 +48,9 @@ gboolean on_request_authorization(Device *device)
     return TRUE;
 }
 
-gboolean callback(gpointer data)
-{
-    GList *devices = binc_adapter_get_devices(default_adapter);
-    for (GList *iter = devices; iter; iter = iter->next)
-    {
-        Device *dev = iter->data;
-        if (dev)
-        {
-            binc_device_disconnect(dev);
-        }
-    }
-    g_list_free(devices);
-
-    if (app != NULL)
-    {
-        binc_adapter_unregister_application(default_adapter, app);
-        binc_application_free(app);
-        app = NULL;
-    }
-
-    if (advertisement != NULL)
-    {
-        binc_adapter_stop_advertising(default_adapter, advertisement);
-        binc_advertisement_free(advertisement);
-    }
-
-    if (agent != NULL)
-    {
-        binc_agent_free(agent);
-    }
-
-    if (default_adapter != NULL)
-    {
-        binc_adapter_free(default_adapter);
-        default_adapter = NULL;
-    }
-
-    g_main_loop_quit((GMainLoop *)data);
-    return FALSE;
-}
-
-static void cleanup_handler(int signo)
-{
-    if (signo == SIGINT)
-    {
-        callback(loop);
-    }
-}
-
 void bt_agent_init()
 {
-    GDBusConnection *dbusconn = g_bus_get_sync(G_BUS_TYPE_SYSTEM, NULL, NULL);
-
-    signal(SIGINT, cleanup_handler);
-
-    loop = g_main_loop_new(NULL, false);
+    dbusconn = g_bus_get_sync(G_BUS_TYPE_SYSTEM, NULL, NULL);
 
     default_adapter = binc_adapter_get_default(dbusconn);
 
@@ -131,15 +80,59 @@ void bt_agent_init()
         binc_application_add_service(app, AVCRP_PROF_UUID);
     }
 
-    g_main_loop_run(loop);
+    pthread_create(&agent_thread_id, NULL, agent_routine, NULL);
+}
 
-    g_main_loop_unref(loop);
+void bt_agent_cleanup()
+{
+    GList *devices = binc_adapter_get_devices(default_adapter);
+    for (GList *iter = devices; iter; iter = iter->next)
+    {
+        Device *dev = iter->data;
+        if (dev)
+        {
+            binc_device_disconnect(dev);
+        }
+    }
+    g_list_free(devices);
+
+    if (app != NULL)
+    {
+        binc_adapter_unregister_application(default_adapter, app);
+        binc_application_free(app);
+        app = NULL;
+    }
+
+    if (agent != NULL)
+    {
+        binc_agent_free(agent);
+        agent = NULL;
+    }
+
+    if (default_adapter != NULL)
+    {
+        binc_adapter_free(default_adapter);
+        default_adapter = NULL;
+    }
+
+    g_main_loop_quit(main_loop);
+
+    pthread_join(agent_thread_id, NULL);
 
     g_dbus_connection_close_sync(dbusconn, NULL, NULL);
     g_object_unref(dbusconn);
 }
 
-void bt_agent_cleanup()
+static void *agent_routine(void *arg)
 {
-    callback(loop);
+    (void)arg;
+
+    GMainContext *main_context = g_main_context_get_thread_default();
+    main_loop = g_main_loop_new(main_context, FALSE);
+
+    g_main_loop_run(main_loop);
+
+    g_main_loop_unref(main_loop);
+
+    return NULL;
 }
